pyproject.toml
[tool.poetry]
name = "discord_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
aiosql = "^9.0"
black = "^23.7.0"

[tool.black]
line-length = 120

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
